requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
autopep8==2.3.1
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.6
passlib==1.7.4
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-c==3.2.1
psycopg2==2.9.9
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.37.2
style==1.1.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
update==0.0.1
uvicorn==0.30.3
watchfiles==0.22.0
websockets==12.0
